* master:
  Add support for updating Container fields as well
  Preserve the raw log fields when wiping using case insensitive contains
  Add soft limit for HTTP Request URI and Header field length. Add a default body_factory template when rejecting a request that's too long
  Doc: Minor typo in CONTRIBUTING.md
  Add QUIC draft-20 support
  Initialize EventIO
  fixed datatype in example plugin
  Cleanup debug logs around SSLWriteBuffer
  Remove unnecesary function name on debug logs in SSLNetVConnection
  TS API - Adds the TSHttpTxnNextHopNameGet() function.

 Conflicts:
	.gitignore
	iocore/net/P_QUICNetVConnection.h
	iocore/net/QUICNetVConnection.cc
	iocore/net/QUICPacketHandler.cc
	iocore/net/quic/Makefile.am
	iocore/net/quic/Mock.h
	iocore/net/quic/QUICAckFrameCreator.cc
	iocore/net/quic/QUICAckFrameCreator.h
	iocore/net/quic/QUICAltConnectionManager.cc
	iocore/net/quic/QUICAltConnectionManager.h
	iocore/net/quic/QUICBidirectionalStream.cc
	iocore/net/quic/QUICBidirectionalStream.h
	iocore/net/quic/QUICConfig.h
	iocore/net/quic/QUICCongestionController.cc
	iocore/net/quic/QUICCryptoStream.cc
	iocore/net/quic/QUICCryptoStream.h
	iocore/net/quic/QUICDebugNames.cc
	iocore/net/quic/QUICFlowController.cc
	iocore/net/quic/QUICFlowController.h
	iocore/net/quic/QUICFrame.cc
	iocore/net/quic/QUICFrame.h
	iocore/net/quic/QUICFrameDispatcher.cc
	iocore/net/quic/QUICFrameDispatcher.h
	iocore/net/quic/QUICFrameGenerator.cc
	iocore/net/quic/QUICFrameGenerator.h
	iocore/net/quic/QUICHandshake.cc
	iocore/net/quic/QUICHandshake.h
	iocore/net/quic/QUICKeyGenerator.cc
	iocore/net/quic/QUICLossDetector.cc
	iocore/net/quic/QUICLossDetector.h
	iocore/net/quic/QUICPacket.cc
	iocore/net/quic/QUICPacket.h
	iocore/net/quic/QUICPacketFactory.cc
	iocore/net/quic/QUICPacketFactory.h
	iocore/net/quic/QUICPacketHeaderProtector.cc
	iocore/net/quic/QUICPacketReceiveQueue.cc
	iocore/net/quic/QUICPacketReceiveQueue.h
	iocore/net/quic/QUICPathValidator.cc
	iocore/net/quic/QUICPathValidator.h
	iocore/net/quic/QUICPinger.cc
	iocore/net/quic/QUICPinger.h
	iocore/net/quic/QUICStreamManager.cc
	iocore/net/quic/QUICStreamManager.h
	iocore/net/quic/QUICTransportParameters.h
	iocore/net/quic/QUICTypes.cc
	iocore/net/quic/QUICTypes.h
	iocore/net/quic/QUICUnidirectionalStream.cc
	iocore/net/quic/QUICUnidirectionalStream.h
	iocore/net/quic/test/test_QUICAckFrameCreator.cc
	iocore/net/quic/test/test_QUICFlowController.cc
	iocore/net/quic/test/test_QUICFrame.cc
	iocore/net/quic/test/test_QUICFrameDispatcher.cc
	iocore/net/quic/test/test_QUICInvariants.cc
	iocore/net/quic/test/test_QUICKeyGenerator.cc
	iocore/net/quic/test/test_QUICLossDetector.cc
	iocore/net/quic/test/test_QUICPacket.cc
	iocore/net/quic/test/test_QUICPacketFactory.cc
	iocore/net/quic/test/test_QUICPacketHeaderProtector.cc
	iocore/net/quic/test/test_QUICStream.cc
	iocore/net/quic/test/test_QUICStreamManager.cc
	iocore/net/quic/test/test_QUICType.cc
	mgmt/RecordsConfig.cc
	src/traffic_quic/quic_client.cc
	src/tscore/ink_inet.cc

+++ example/plugins/c-api/server_transform/server_transform.c
@@ -284,7 +284,7 @@ static int
 transform_buffer_event(TSCont contp, TransformData *data, TSEvent event ATS_UNUSED, void *edata ATS_UNUSED)
 {
   TSVIO write_vio;
-  int towrite;
+  int64_t towrite;
 
   if (!data->input_buf) {
     data->input_buf    = TSIOBufferCreate();
@@ -313,7 +313,7 @@ transform_buffer_event(TSCont contp, TransformData *data, TSEvent event ATS_UNUS
   if (towrite > 0) {
     /* The amount of data left to read needs to be truncated by
        the amount of data actually in the read buffer. */
-    int avail = TSIOBufferReaderAvail(TSVIOReaderGet(write_vio));
+    int64_t avail = TSIOBufferReaderAvail(TSVIOReaderGet(write_vio));
     if (towrite > avail) {
       towrite = avail;
     }
